[compiler] Custom type definitions in config
Summary:
This PR allows custom type definitions for globals to be loaded in the Forget config. We use a slightly different representation than we do internally, in order to ensure that things are well formed before we pass them to the shape registry. The most fundamental issue is that we introduce a notion of IDs (expected to be unique) for types: we can introduce a type and then refer to that type later by a unique ID in a function return type if we have, for example, a global value that returns itself when called.

ghstack-source-id: 46791e48af2410996974747cc4849acfa63fdc20
Pull Request resolved: https://github.com/facebook/react/pull/30670