[Flight][Static] When prerendering serialize infinite promise when aborting with no reason
When prerendering if you abort the prerender without a reason instead of erroring each remaining task complete it with a promise that never resolve

Unfortunately when you abort with an AbortSignal without a value the aborted reason is defaulted to an AbortError DOMexception. We test for this and basically just say that if you abort with an AbortError DOMException that is equivalent to aborting with nothing. Practically this should be fine because usually you abort with a specific reason.